core: Centralize notification triggers in to notification redux slice
